@W-14584326: [Android] OTPs Sample App (Update Login Activity Structure To Improve Readability By Moving More Functions Into Activity And Navigation Into View Model)
